+++ test/dartemis/core/components_setup.dart
@@ -4,7 +4,7 @@ import "package:dartemis/dartemis.dart";
 
 final Type COMPONENT_A = new ComponentA().runtimeType;
 final Type COMPONENT_B = new ComponentB().runtimeType;
-final Type COMPONENT_C = new ComponentPoolableC().runtimeType;
+final Type COMPONENT_C = new PooledComponentC().runtimeType;
 
 const int COMPONENT_A_BIT = 0x0001;
 const int COMPONENT_B_BIT = 0x0002;
@@ -18,7 +18,7 @@ void setUpComponents() {
 
 class ComponentA extends Component {}
 class ComponentB extends Component {}
-class ComponentPoolableC extends ComponentPoolable {
-  ComponentPoolableC._();
-  factory ComponentPoolableC() => new Poolable.of(ComponentPoolableC, () => new ComponentPoolableC._());
+class PooledComponentC extends PooledComponent {
+  PooledComponentC._();
+  factory PooledComponentC() => new Pooled.of(PooledComponentC, () => new PooledComponentC._());
 }